+++ packages/webview_flutter/webview_flutter_android/lib/webview_surface_android.dart
@@ -17,12 +17,6 @@ import 'webview_android_widget.dart';
 /// Android [WebViewPlatform] that uses [AndroidViewSurface] to build the [WebView] widget.
 ///
 /// To use this, set [WebView.platform] to an instance of this class.
-///
-/// This implementation uses hybrid composition to render the [WebView] on
-/// Android. It solves multiple issues related to accessibility and interaction
-/// with the [WebView] at the cost of some performance on Android versions below
-/// 10. See https://github.com/flutter/flutter/wiki/Hybrid-Composition for more
-/// information.
 class SurfaceAndroidWebView extends AndroidWebView {
   @override
   Widget build({